Removed sign extension function, used type casting instead.

+++ immediates.h
@@ -1,7 +1,9 @@
 #ifndef IMMEDIATES_H
 #define IMMEDIATES_H
 
-unsigned long long get_displacement(unsigned char * inst, int op_byte_num, int disp_len, int instr_len);
+#include <stdlib.h>
+
+int64_t get_displacement(unsigned char * inst, int op_byte_num, int disp_len, int instr_len);
 unsigned long long get_immediate(unsigned char * inst, int op_byte_num, int imm_len);
 
 #endif
